mv chat user module
