+++ src/event-logger/event-logger.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { SlackMessageClient } from 'src/api/slack/slack-api';
 import { ZapierWebhookClient } from 'src/api/zapier/zapier-webhook-client';
-import { ChatUserEntity } from 'src/entities/chat-user.entity';
+import { ChatUserModule } from 'src/chat-user/chat-user.module';
 import { EventLogEntity } from 'src/entities/event-log.entity';
 import { PartnerAccessEntity } from 'src/entities/partner-access.entity';
 import { PartnerEntity } from 'src/entities/partner.entity';
@@ -23,7 +23,6 @@ import { EventLoggerService } from './event-logger.service';
 @Module({
   imports: [
     TypeOrmModule.forFeature([
-      ChatUserEntity,
       EventLogEntity,
       UserEntity,
       PartnerAccessEntity,
@@ -32,6 +31,7 @@ import { EventLoggerService } from './event-logger.service';
       TherapySessionEntity,
       SubscriptionEntity,
     ]),
+    ChatUserModule,
   ],
   controllers: [EventLoggerController],
   providers: [
