+++ src/front-chat/front-chat.scheduler.spec.ts
@@ -1,9 +1,9 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { MAILCHIMP_CUSTOM_EVENTS } from 'src/api/mailchimp/mailchimp-api.interfaces';
+import { ChatUserService } from 'src/chat-user/chat-user.service';
 import { ChatUserEntity } from 'src/entities/chat-user.entity';
 import { ServiceUserProfilesService } from 'src/service-user-profiles/service-user-profiles.service';
 import { FrontChatScheduler } from './front-chat.scheduler';
-import { FrontChatService } from './front-chat.service';
 
 // End-to-end behaviour of the unread-message notification flow: the cron fires,
 // the service yields eligible users, the scheduler marks pending, hands the event
@@ -11,12 +11,13 @@ import { FrontChatService } from './front-chat.service';
 
 describe('FrontChatScheduler — unread notification flow', () => {
   let scheduler: FrontChatScheduler;
-  let frontChatService: jest.Mocked<Pick<
-    FrontChatService,
+  let chatUserService: jest.Mocked<Pick<
+    ChatUserService,
     | 'getUsersWithUnreadMessages'
     | 'markUnreadNotificationPending'
     | 'markUnreadNotificationSent'
     | 'markUnreadNotificationFailed'
+    | 'recoverStalePendingNotifications'
   >>;
   let serviceUserProfiles: jest.Mocked<Pick<ServiceUserProfilesService, 'sendMailchimpUserEventWithRecovery'>>;
 
@@ -24,11 +25,12 @@ describe('FrontChatScheduler — unread notification flow', () => {
     ({ id, userId: `u-${id}`, unreadNotificationAttempts: 0 }) as ChatUserEntity;
 
   beforeEach(async () => {
-    frontChatService = {
+    chatUserService = {
       getUsersWithUnreadMessages: jest.fn().mockResolvedValue([]),
       markUnreadNotificationPending: jest.fn().mockResolvedValue(undefined),
       markUnreadNotificationSent: jest.fn().mockResolvedValue(undefined),
       markUnreadNotificationFailed: jest.fn().mockResolvedValue(undefined),
+      recoverStalePendingNotifications: jest.fn().mockResolvedValue(0),
     };
     serviceUserProfiles = {
       sendMailchimpUserEventWithRecovery: jest.fn().mockResolvedValue(undefined),
@@ -37,7 +39,7 @@ describe('FrontChatScheduler — unread notification flow', () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         FrontChatScheduler,
-        { provide: FrontChatService, useValue: frontChatService },
+        { provide: ChatUserService, useValue: chatUserService },
         { provide: ServiceUserProfilesService, useValue: serviceUserProfiles },
       ],
     }).compile();
@@ -46,15 +48,15 @@ describe('FrontChatScheduler — unread notification flow', () => {
   });
 
   it('marks pending → sends Mailchimp event → marks sent for each eligible user', async () => {
-    frontChatService.getUsersWithUnreadMessages.mockResolvedValue([
+    chatUserService.getUsersWithUnreadMessages.mockResolvedValue([
       { chatUser: buildChatUser('cu-1'), email: 'a@example.com' },
       { chatUser: buildChatUser('cu-2'), email: 'b@example.com' },
     ]);
 
     await scheduler.checkUnreadMessages();
 
-    expect(frontChatService.markUnreadNotificationPending).toHaveBeenCalledWith('cu-1');
-    expect(frontChatService.markUnreadNotificationPending).toHaveBeenCalledWith('cu-2');
+    expect(chatUserService.markUnreadNotificationPending).toHaveBeenCalledWith('cu-1');
+    expect(chatUserService.markUnreadNotificationPending).toHaveBeenCalledWith('cu-2');
     expect(serviceUserProfiles.sendMailchimpUserEventWithRecovery).toHaveBeenCalledWith(
       'a@example.com',
       MAILCHIMP_CUSTOM_EVENTS.FRONT_MESSAGE_UNREAD,
@@ -63,23 +65,23 @@ describe('FrontChatScheduler — unread notification flow', () => {
       'b@example.com',
       MAILCHIMP_CUSTOM_EVENTS.FRONT_MESSAGE_UNREAD,
     );
-    expect(frontChatService.markUnreadNotificationSent).toHaveBeenCalledWith('cu-1');
-    expect(frontChatService.markUnreadNotificationSent).toHaveBeenCalledWith('cu-2');
-    expect(frontChatService.markUnreadNotificationFailed).not.toHaveBeenCalled();
+    expect(chatUserService.markUnreadNotificationSent).toHaveBeenCalledWith('cu-1');
+    expect(chatUserService.markUnreadNotificationSent).toHaveBeenCalledWith('cu-2');
+    expect(chatUserService.markUnreadNotificationFailed).not.toHaveBeenCalled();
   });
 
   it('marks pending before the Mailchimp call so a restart between the two cannot double-send', async () => {
     const order: string[] = [];
-    frontChatService.getUsersWithUnreadMessages.mockResolvedValue([
+    chatUserService.getUsersWithUnreadMessages.mockResolvedValue([
       { chatUser: buildChatUser('cu-1'), email: 'a@example.com' },
     ]);
-    frontChatService.markUnreadNotificationPending.mockImplementation(async () => {
+    chatUserService.markUnreadNotificationPending.mockImplementation(async () => {
       order.push('pending');
     });
     serviceUserProfiles.sendMailchimpUserEventWithRecovery.mockImplementation(async () => {
       order.push('mailchimp');
     });
-    frontChatService.markUnreadNotificationSent.mockImplementation(async () => {
+    chatUserService.markUnreadNotificationSent.mockImplementation(async () => {
       order.push('sent');
     });
 
@@ -89,7 +91,7 @@ describe('FrontChatScheduler — unread notification flow', () => {
   });
 
   it('records the failure and does not mark sent when the Mailchimp event throws', async () => {
-    frontChatService.getUsersWithUnreadMessages.mockResolvedValue([
+    chatUserService.getUsersWithUnreadMessages.mockResolvedValue([
       { chatUser: buildChatUser('cu-1'), email: 'a@example.com' },
     ]);
     serviceUserProfiles.sendMailchimpUserEventWithRecovery.mockRejectedValueOnce(
@@ -98,16 +100,16 @@ describe('FrontChatScheduler — unread notification flow', () => {
 
     await scheduler.checkUnreadMessages();
 
-    expect(frontChatService.markUnreadNotificationPending).toHaveBeenCalledWith('cu-1');
-    expect(frontChatService.markUnreadNotificationSent).not.toHaveBeenCalled();
-    expect(frontChatService.markUnreadNotificationFailed).toHaveBeenCalledWith(
+    expect(chatUserService.markUnreadNotificationPending).toHaveBeenCalledWith('cu-1');
+    expect(chatUserService.markUnreadNotificationSent).not.toHaveBeenCalled();
+    expect(chatUserService.markUnreadNotificationFailed).toHaveBeenCalledWith(
       'cu-1',
       'mailchimp 502',
     );
   });
 
   it('isolates failures per user — one failing send does not block the others in the batch', async () => {
-    frontChatService.getUsersWithUnreadMessages.mockResolvedValue([
+    chatUserService.getUsersWithUnreadMessages.mockResolvedValue([
       { chatUser: buildChatUser('cu-1'), email: 'a@example.com' },
       { chatUser: buildChatUser('cu-2'), email: 'b@example.com' },
       { chatUser: buildChatUser('cu-3'), email: 'c@example.com' },
@@ -118,23 +120,22 @@ describe('FrontChatScheduler — unread notification flow', () => {
 
     await scheduler.checkUnreadMessages();
 
-    expect(frontChatService.markUnreadNotificationSent).toHaveBeenCalledWith('cu-1');
-    expect(frontChatService.markUnreadNotificationFailed).toHaveBeenCalledWith('cu-2', 'boom');
-    expect(frontChatService.markUnreadNotificationSent).toHaveBeenCalledWith('cu-3');
-    expect(frontChatService.markUnreadNotificationSent).not.toHaveBeenCalledWith('cu-2');
+    expect(chatUserService.markUnreadNotificationSent).toHaveBeenCalledWith('cu-1');
+    expect(chatUserService.markUnreadNotificationFailed).toHaveBeenCalledWith('cu-2', 'boom');
+    expect(chatUserService.markUnreadNotificationSent).toHaveBeenCalledWith('cu-3');
+    expect(chatUserService.markUnreadNotificationSent).not.toHaveBeenCalledWith('cu-2');
   });
 
   it('skips the Mailchimp send entirely if the pending DB write fails, so we never send unrecorded', async () => {
-    frontChatService.getUsersWithUnreadMessages.mockResolvedValue([
+    chatUserService.getUsersWithUnreadMessages.mockResolvedValue([
       { chatUser: buildChatUser('cu-1'), email: 'a@example.com' },
     ]);
-    frontChatService.markUnreadNotificationPending.mockRejectedValueOnce(new Error('db down'));
+    chatUserService.markUnreadNotificationPending.mockRejectedValueOnce(new Error('db down'));
 
     await scheduler.checkUnreadMessages();
 
     expect(serviceUserProfiles.sendMailchimpUserEventWithRecovery).not.toHaveBeenCalled();
-    expect(frontChatService.markUnreadNotificationSent).not.toHaveBeenCalled();
-    expect(frontChatService.markUnreadNotificationFailed).not.toHaveBeenCalled();
+    expect(chatUserService.markUnreadNotificationSent).not.toHaveBeenCalled();
+    expect(chatUserService.markUnreadNotificationFailed).not.toHaveBeenCalled();
   });
-
 });
