+++ src/partner-access/partner-access.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { SlackMessageClient } from 'src/api/slack/slack-api';
 import { ZapierWebhookClient } from 'src/api/zapier/zapier-webhook-client';
 import { CourseUserService } from 'src/course-user/course-user.service';
-import { ChatUserEntity } from 'src/entities/chat-user.entity';
+import { ChatUserModule } from 'src/chat-user/chat-user.module';
 import { CourseUserEntity } from 'src/entities/course-user.entity';
 import { EventLogEntity } from 'src/entities/event-log.entity';
 import { PartnerAccessEntity } from 'src/entities/partner-access.entity';
@@ -29,7 +29,6 @@ import { PartnerAccessService } from './partner-access.service';
 @Module({
   imports: [
     TypeOrmModule.forFeature([
-      ChatUserEntity,
       UserEntity,
       PartnerEntity,
       PartnerAccessEntity,
@@ -40,6 +39,7 @@ import { PartnerAccessService } from './partner-access.service';
       SubscriptionEntity,
       EventLogEntity,
     ]),
+    ChatUserModule,
     FirebaseModule,
   ],
   controllers: [PartnerAccessController],
