+++ src/resource-user/resource-user.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { SlackMessageClient } from 'src/api/slack/slack-api';
 import { ZapierWebhookClient } from 'src/api/zapier/zapier-webhook-client';
-import { ChatUserEntity } from 'src/entities/chat-user.entity';
+import { ChatUserModule } from 'src/chat-user/chat-user.module';
 import { CourseEntity } from 'src/entities/course.entity';
 import { EventLogEntity } from 'src/entities/event-log.entity';
 import { PartnerAccessEntity } from 'src/entities/partner-access.entity';
@@ -31,7 +31,6 @@ import { ResourceUserService } from './resource-user.service';
 @Module({
   imports: [
     TypeOrmModule.forFeature([
-      ChatUserEntity,
       UserEntity,
       ResourceEntity,
       ResourceUserEntity,
@@ -46,6 +45,7 @@ import { ResourceUserService } from './resource-user.service';
       SubscriptionEntity,
       EventLogEntity,
     ]),
+    ChatUserModule,
   ],
   controllers: [ResourceUserController],
   providers: [
