+++ src/therapy-session/therapy-session.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { SlackMessageClient } from 'src/api/slack/slack-api';
 import { ZapierWebhookClient } from 'src/api/zapier/zapier-webhook-client';
-import { ChatUserEntity } from 'src/entities/chat-user.entity';
+import { ChatUserModule } from 'src/chat-user/chat-user.module';
 import { EventLogEntity } from 'src/entities/event-log.entity';
 import { PartnerAccessEntity } from 'src/entities/partner-access.entity';
 import { PartnerEntity } from 'src/entities/partner.entity';
@@ -23,7 +23,6 @@ import { TherapySessionService } from './therapy-session.service';
 @Module({
   imports: [
     TypeOrmModule.forFeature([
-      ChatUserEntity,
       TherapySessionEntity,
       UserEntity,
       PartnerAccessEntity,
@@ -32,6 +31,7 @@ import { TherapySessionService } from './therapy-session.service';
       EventLogEntity,
       SubscriptionEntity,
     ]),
+    ChatUserModule,
   ],
   controllers: [TherapySessionController],
   providers: [
